[Feat] #1 MainViewController 기초 구성
